requirements.txt
httpx>=0.27
PyYAML>=6.0
feedparser>=6.0
lxml>=5.2
selectolax>=0.3.21
python-dateutil>=2.9
orjson>=3.10
playwright>=1.45
